#1 refactor: Ui 폴더 구조 리패키징
