+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Zenefit_Android"
         tools:targetApi="31">
-        <activity android:name=".SplashActivity"
+        <activity android:name=".presentation.ui.splash.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.Zenefit_Android.Splash">
             <intent-filter>
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="true">
 
         </activity>
