+++ app/src/main/java/com/zenefit/zenefit_android/presentation/ui/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.zenefit.zenefit_android
+package com.zenefit.zenefit_android.presentation.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.zenefit.zenefit_android.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
